+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Version 2.0.beta work in progress
 - new #107 Added convenient method of adding link to the object (VladimirShestakov)
 - enh #151 Attribute is_delete is renamed to is_draft, Entity->isDelete() to Entity->isDragt() (VladimirShestakov)
 - enh #152 By default all new objects are drafts (VladimirShestakov)
+- new #154 The table "text" is used for values with a length more than 255 characters (VladimirShestakov)
 
 Version 2.0.alpha September, 30 2013
 ----------------------------